move authApiService
