+++ app/javascript/authentication/context/UserProvider.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useReducer } from "react"
 
-import { getProfile, signIn } from "../authApiService"
+import { getProfile, signIn } from "../../api/authApiService"
 import { isTokenValid } from "../token"
 import { saveProfile, signOut, startLoading, stopLoading } from "./userActions"
 import UserContext, { ContextProps } from "./UserContext"
